requirements.txt
langchain
langchain-community
langchain-openai
huggingface_hub
Pillow